==> step1_initial_useful_cols.ipynb <==
# Importing Necessary Libraries
# =============================

import pyspark
from pyspark.sql.functions import col
from pyspark.sql import functions as F, SparkSession
from pyspark.sql.types import FloatType

# Reading Training from Tables into DataFrames
# ==========================================================
# Prior to running this code cell, the training data must be uploaded onto
# DataBricks as a table named "training_data".

# Training Data
df = sqlContext.table("training_data")
df.printSchema()

# Dataframe size
# =============================

row_number = df.count()
column_number = len(df.dtypes)
print('The Train data has ', row_number, 'individual reddits posts (rows) and', column_number, 'associated attributes (columns)')

# Plot count of each data type.
# =============================

dataname = []
datatype = []
for name, dtype in df.dtypes:
  if dtype not in ['boolean', 'string', 'bigint', 'array<string>']:
    dataname.append(name)
    datatype.append('other')
  else:
    dataname.append(name)
    datatype.append(dtype)
  
datanamedatatype = sqlContext.createDataFrame(zip(dataname, datatype), schema=['dataname', 'datatype'])
# x.show(100)
display(datanamedatatype.groupby('datatype').count())

# # Some data exploration of each column.
# # =============================

# for item in dataname:
#   df.select(item, 'score', 'id').groupby(item).agg({'id':'count', 'score':'average'}).show(5)

# Columns of each data type.
# =============================

datanamedatatype.createOrReplaceTempView("datanamedatatype")

sql_results_string = spark.sql("SELECT dataname FROM datanamedatatype WHERE datatype = 'string'")
sql_results_bigint = spark.sql("SELECT dataname FROM datanamedatatype WHERE datatype = 'bigint'")
sql_results_boolean = spark.sql("SELECT dataname FROM datanamedatatype WHERE datatype = 'boolean'")

string_cols = sql_results_string.select("dataname").rdd.flatMap(lambda x: x).collect()
bigint_cols = sql_results_bigint.select("dataname").rdd.flatMap(lambda x: x).collect()
boolean_cols = sql_results_boolean.select("dataname").rdd.flatMap(lambda x: x).collect()

# Function for finding Columns with Missing Values/Null values.
# =============================

def missingcounts(df):
  dfMissing = df.select([F.count(F.when(F.isnan(i) | \
                                   F.col(i).contains('NA') | \
                                   F.col(i).contains('NULL') | \
                                   F.col(i).isNull(), i)).alias(i) \
                    for i in df.columns])
  return dfMissing

# Null count for String cols.
# =============================

dfstring = df.select([c for c in df.columns if c in string_cols])
dfstringMissing = missingcounts(dfstring)

NullCount = []
for i in string_cols:
  tableList = dfstringMissing.select(i).rdd.flatMap(lambda x: x).collect()
  NullCount.append(tableList[0])

dfstringCountNull = sqlContext.createDataFrame(zip(string_cols, NullCount), schema=['String Columns', 'Null/Missing Count'])

display(dfstringCountNull)

# Null count for Bigint cols.
# =============================

dfbigint = df.select([c for c in df.columns if c in bigint_cols])
dfbigintMissing = missingcounts(dfbigint)

NullCount = []
for i in bigint_cols:
  tableList = dfbigintMissing.select(i).rdd.flatMap(lambda x: x).collect()
  NullCount.append(tableList[0])

dfbigintCountNull = sqlContext.createDataFrame(zip(bigint_cols, NullCount), schema=['Bigint Columns', 'Null/Missing Counts'])

display(dfbigintCountNull)

# Determing useful columns intuitively - quick data clean.
# =============================

print("Total columns:", len(dataname))

# keep only relevant string cols with low null values or intuitive string columns.
drop_string_cols = []

keep_str_cols = ["author", "domain", "permalink", "subreddit_id", "suggested_sort", "title", "whitelist_status"]
for stringc in string_cols:
  if stringc not in keep_str_cols:
    drop_string_cols.append(stringc)    
    
for item in drop_string_cols:
  if item in dataname:
    dataname.remove(item)

print('Total columns (after removing null/irrelavent string cols):', len(dataname), ' - dropped ', len(drop_string_cols), ' string out of ', len(string_cols), '(includes permalink, author and title that are used for deriving other features but not used as features themselves)')

# drop all bigint_cols with null values or with same values.
drop_bigint_cols = ['thumbnail_height', 'thumbnail_width', 'retrieved_on']

countr = 0 
for bigintc in bigint_cols:
  x = dfbigint.select(bigintc).distinct().count()
  if x <= 1:
    drop_bigint_cols.append(bigintc)
    countr += 1

for item in drop_bigint_cols:
  if item in dataname:
    dataname.remove(item)
    
print('Total columns (after removing 3 null bigint cols and ', countr,  ' bigint cols with same values):', len(dataname), ' - dropped ', len(drop_bigint_cols), ' bigint out of ', len(bigint_cols), ('includes permalink, author and title that are used for deriving other features but not used as features themselves'))

# drop boolean columns that have same values
drop_boolean_cols = []

# dfboolean = df.select([col(c).cast('double') for c in df.columns if c in boolean_cols])
dfboolean = df.select([c for c in df.columns if c in boolean_cols])

for boolc in boolean_cols:
  x = dfboolean.select(boolc).distinct().count()
  if x <= 1:
    drop_boolean_cols.append(boolc)
    
for item in drop_boolean_cols:
  if item in dataname:
    dataname.remove(item)
    
print('Total columns (after removing all same value):', len(dataname), ' - dropped ', len(drop_boolean_cols), ' boolean out of ', len(boolean_cols))

# drop all other / array columns.
drop_othernarray_cols = ['author_flair_richtext', 'link_flair_richtext', 'preview']

for item in drop_othernarray_cols:
  if item in dataname:
    dataname.remove(item)

print('total columns (after dropping other/array cols):', len(dataname), ' - dropped all 3')

# Printing useful columns  - to be used for next step.
# =============================

print('The useful columns are:', dataname)
